Relax ownership requirements on Cached<K,V>
This change relaxes the requirement of &K on get-style functions
to accept a &Q that K can be borrowed as, and that meets the minimum
reasonable needs of the Cached trait.

It follows the model of the Rust standard HashMap. Each function is
generic over the key type.

This enables you to have caches with String keys, while querying
with only an &str. This reduces copies in user code, and helps real-
world use cases pay as little as possible for each cache hit.

In practice, this reduces the auth cache latency in my service's
frontend from 387ns to 338ns (for hits) by avoiding the string copy
for the authorization header. While the absolute amount per
invocation is not much, this reduces the memory pressure on every
api call we service.

Downstream implementations of Cached<K,V> will need to update their
bounds on cache_get, cache_get_mut and cache_remove. User code that
works with the current Cached trait should continue to work with
the new bounds, as it is a relaxing of the constraint. Only folks
who are directly integrated with the trait will need to update, and
it should just make their apps slightly quicker or at worst no slower.

I've taken the liberty of adding a couple simple doctests for the
Cached trait and the functions I've adjusted to demonstrate their
use and to ensure they keep working that way.

I've left the macros alone for now. I don't use them anymore, and
I'm not very familiar with them. It was not readily apparent to me
how to take advantage of borrowable trait copy-avoidance in them,
but I'm sure someone else will be able to add it!